* commit '45929938c7e2cf9d4077638941d3fe1ec28400c7':
  libserver: handle duplicate named property requests
  libserver: avoid returning high-numbered named property ids from NamedPropertyMapper
  libserver: return lowest named property id on getIDsFromNames
  libserver: avoid returning high-numbered named property ids from getIDsFromNames
  client: reject high-numbered named property ids from server
  libserver: turn SQL field names.guid into a binary(16)
  libserver: set indexes on "names" SQL table to be unique

+++ provider/client/ECNamedProp.h
@@ -38,7 +38,7 @@ public:
 
 private:
 	KCHL::object_ptr<WSTransport> lpTransport;
-	std::map<MAPINAMEID *,ULONG,ltmap>		mapNames;
+	std::map<MAPINAMEID *, unsigned int, ltmap> mapNames;
 
 	HRESULT			ResolveLocal(MAPINAMEID *lpName, ULONG *ulId);
 	HRESULT			ResolveCache(MAPINAMEID *lpName, ULONG *ulId);
